Added first version of RateLimiter service
First of all, as per the storage, the interface is there in order to be able to swap for different types
of ratelimiters (maybe, in the future, it could be based on tokens, other types of auth ecc)
without much hassle.

Then, for the concrete implementation itself:
the algorithm is quite simple, it just saves timestamps of visits and compares them to current time of request;
after evicting expired timestamps from the storage, it adds current time and then compares the number of
remaining timestamps with the limit of the current type of request (only POST|GET are allowed).

Value are saved on a per-ip per-method basis, so that a user can make the maximum number of requests of
each type during the same time window.

The time window is moving, meaning that it always counts requests made in the last 60 seconds,
regardless of absolute time.

Parameters are all configurable through environment variables, therefore in the future a system which has
custom rates and time windows for each user (or maybe group of users) can be easily implemented.

There is also a simple whitelist. This is currently implemented through env variable as well, but probably
it would be better suited as a database item.

Furthermore, Symfony's Request->getClientIp() automatically handles 'X_FORWARDED_FOR' and
'HTTP_X_FORWARDER_FOR' headers, which can be present if the request travelled through some proxies.
In the future further configuration is possible to refine this behaviour, maybe blocking known malicious proxies
while allowing trusted ones.