Added frontend page to test functionalities
This page is used to test functionalities of this app, mainly to make it easier to test POST requests
without using tools like POSTMAN and such.

Since the code is very simple and the implementation goes beyond the scope of this exercise, I opted not to install
twig bundle for templating or other libraries for js or style.